rexstan enhancements

+++ lib/category.php
@@ -257,7 +257,7 @@ class Category implements \D2U_Helper\ITranslationHelper
             $parameterArray = [];
             $parameterArray['category_id'] = $this->category_id;
 
-            $this->url = rex_getUrl(rex_config::get('d2u_immo', 'article_id'), $this->clang_id, $parameterArray, '&');
+            $this->url = rex_getUrl((int) rex_config::get('d2u_immo', 'article_id'), $this->clang_id, $parameterArray, '&');
         }
 
         if ($including_domain) {
